Minor changes
- Change default max cpu
- Change results' directory organization
-  Beautify prints

+++ src/vqe.py
@@ -79,7 +79,7 @@ class VQE:
         return f"""\nVQE instance 
         Ansatz:
             qubits: {self.ansatz.num_qubits}
-            layers: {self.ansatz.num_parameters / self.ansatz.num_qubits - 1}
+            layers: {int(self.ansatz.num_parameters / self.ansatz.num_qubits - 1)}
             parameters: {self.ansatz.num_parameters}
         Optimizer: {self.optimizer}
         Simulator: {self.simulator}
